Correction of the lock-free enqueue: swap the tail pointer first, before setting the old tail's next pointer.
__sync_swap() would be a good thing here.

I think there is a conceivable race condition here:
When there is a single element in the queue, head equals tail.
If an enqueue races with a dequeue in that condition, head could be set to nil by the dequeue() before the old head's next pointer got set by the enqueue()

In the end, it is probably not possible to implement this queue in a lock-free way, given the limited set of atomic operations usable from Swift.
(OSAtomicFifoEnqueue and OSAtomicFifoDequeue obviously exist; do they depend on x64 instructions not implemented by ARM?)

The spin lock is surprisingly effective; the protected section is short, so the risks of a bad priority inversion situation are minimal.