Acknowledging that Swift messes with lock-free queues
The established algorithms (Michael&Scott, Optimistic) assume memory that is either allocated or not.
Swift uses a 3-state memory that is either unallocated, allocated, or allocated-and-initialized.
Attempting to perform an ARC operation on non-initialized memory results in a crash.

The dequeuing operation (in both algorithms) has a step that consists of optimistically reading the element to be returned prior to the compare-and-swap (CAS) operation.
That can (and does) get into a race condition with the destroy() step in another thread. Such crashes make the algorithms unusable.

(For some reason, they managed to work okay in Swift prior to 2.0)